% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_DEGs}
\alias{example_DEGs}
\title{Example DEGs}
\format{
A data frame with ...5,187 rows and 6 columns
}
\source{
GSE165816
}
\usage{
example_DEGs
}
\description{
A dataset of differentially expressed genes when comparing fibroblasts from healing vs non-healing diabetic foot ulcers.
}
\keyword{datasets}
