% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_GRN}
\alias{example_GRN}
\title{Example Gene regulatory network (GRN)}
\format{
A named list
}
\source{
Inferred using pySCENIC
}
\usage{
example_GRN
}
\description{
Fibroblast GRN.
}
\keyword{datasets}
